add: Listの中身をclient componentsに分割
