+++ src/layouts/List/Received/index.css.ts
@@ -1,0 +1,1 @@
+export * from '../common.css'
