+++ src/layouts/List/index.tsx
@@ -1,14 +1,9 @@
-import * as styles from './index.css'
 import Header from './Header'
-import { FaPen, FaPlus } from 'react-icons/fa6'
-import Card from '../../components/Card'
 import { getCreatedCards } from '../../utils/strapi/card'
 import { redirect } from 'next/navigation'
-import Link from 'next/link'
-import { mediaRecordsToUrlSet } from '../../utils/strapi/strapiImage'
 import { getReceivedCards } from '../../utils/strapi/receivedCard'
-import Image from 'next/image'
-import emptyPost from './empty-post.svg'
+import Created from './Created'
+import Received from './Received'
 
 type ListProps = {
   tab: 'created' | 'received'
@@ -23,125 +18,23 @@ const List = async ({ tab }: ListProps) => {
     return (
       <div>
         <Header activeTab={tab} />
-        <div className={styles.container}>
-          {cards.data.length > 0 ? (
-            <div className={styles.cardContainer}>
-              <Link href="/create/new" className={styles.cardLink}>
-                <div className={styles.newCardButtonSizeInner}>
-                  <div className={styles.newCardButtonContent}>
-                    <div>
-                      <FaPlus className={styles.newCardButtonIcon} />
-                    </div>
-                  </div>
-                </div>
-                <div className={styles.newCardButtonTextContainer}>
-                  <div className={styles.newCardButtonText}>新規作成</div>
-                </div>
-              </Link>
-
-              {cards.data.map((card, index) => (
-                <div className={styles.content} key={index}>
-                  <Link
-                    href={`/create/detail/${card.id}`}
-                    className={styles.cardLink}
-                  >
-                    <div className={styles.card}>
-                      <Card
-                        layout={card.attributes.view.layout}
-                        background={card.attributes.view.background}
-                        userImages={mediaRecordsToUrlSet(
-                          card.attributes.userImages.data
-                        )}
-                        maxFormat="small"
-                      />
-                    </div>
-                    {card.attributes.publishedAt === null ? (
-                      <div className={styles.draftCardTitle}>
-                        <FaPen size={16} />
-                        下書き
-                      </div>
-                    ) : (
-                      <div className={styles.cardTitle}>
-                        {card.attributes.title}
-                      </div>
-                    )}
-                  </Link>
-                </div>
-              ))}
-            </div>
-          ) : (
-            <Link href="/create/new" className={styles.cardLink}>
-              <div className={styles.newCardContainer}>
-                <FaPlus className={styles.newCardIcon} />
-                <div className={styles.newCardText}>年賀状を新規作成</div>
-              </div>
-            </Link>
-          )}
-        </div>
+        <Created initialCards={cards} />
       </div>
     )
   }
   if (tab === 'received') {
     const receivedCards = await getReceivedCards()
     if (!receivedCards) {
-      redirect('/')
+      redirect(
+        `/api/auth/signin?callbackUrl=${encodeURIComponent('/receive/list')}`
+      )
     }
     return (
       <div>
         <Header activeTab={tab} />
-        <div className={styles.container}>
-          {receivedCards.data.length > 0 ? (
-            <div className={styles.cardContainer}>
-              {receivedCards.data.map((receivedCard, index) => {
-                if (
-                  !receivedCard.attributes.card ||
-                  !receivedCard.attributes.card.data
-                ) {
-                  return null;
-                }
-                return (
-                  <div className={styles.content} key={index}>
-                    <Link
-                      href={`/receive/detail/${receivedCard.id}`}
-                      className={styles.cardLink}
-                    >
-                      <div className={styles.card}>
-                        <Card
-                          layout={
-                            receivedCard.attributes.card.data.attributes.view.layout
-                          }
-                          background={
-                            receivedCard.attributes.card.data.attributes.view
-                              .background
-                          }
-                          userImages={mediaRecordsToUrlSet(
-                            receivedCard.attributes.card.data.attributes.userImages
-                              .data
-                          )}
-                          maxFormat="small"
-                        />
-                      </div>
-                      <div className={styles.cardTitle}>
-                        {receivedCard.attributes.card.data.attributes.creatorName}
-                      </div>
-                    </Link>
-                  </div>
-                );
-              })}
-            </div>
-          ) : (
-            <>
-              <div className={styles.noCardContainer}>
-                <div className={styles.noCardMessage}>友達と年賀状を共有しよう！</div>
-              </div>
-              <div className={styles.noCardContainer}>
-                <Image src={emptyPost} alt='emptyPost'></Image>
-              </div>
-            </>
-          )}
-        </div>
+        <Received initialReceivedCards={receivedCards} />
       </div>
-    );
+    )
   }
 }
